Fixed MSVC L4 warnings.

+++ examples/00-helloworld/helloworld.cpp
@@ -47,7 +47,7 @@ class ExampleHelloWorld : public entry::AppI
 		if (!entry::processEvents(m_width, m_height, m_debug, m_reset) )
 		{
 			// Set view 0 default viewport.
-			bgfx::setViewRect(0, 0, 0, m_width, m_height);
+			bgfx::setViewRect(0, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 
 			// This dummy draw call is here to make sure that view 0 is cleared
 			// if no other draw calls are submitted to view 0.
@@ -55,8 +55,8 @@ class ExampleHelloWorld : public entry::AppI
 
 			// Use debug font to print information about this example.
 			bgfx::dbgTextClear();
-			bgfx::dbgTextImage(bx::uint16_max(m_width /2/8, 20)-20
-							 , bx::uint16_max(m_height/2/16, 6)-6
+			bgfx::dbgTextImage(bx::uint16_max(uint16_t(m_width /2/8 ), 20)-20
+							 , bx::uint16_max(uint16_t(m_height/2/16), 6)-6
 							 , 40
 							 , 12
 							 , s_logo
