+++ examples/03-raymarch/raymarch.cpp
@@ -37,7 +37,7 @@ inline void mtxProj(float* _result, float _fovy, float _aspect, float _near, flo
 	bx::mtxProj(_result, _fovy, _aspect, _near, _far, s_oglNdc);
 }
 
-void renderScreenSpaceQuad(uint32_t _view, bgfx::ProgramHandle _program, float _x, float _y, float _width, float _height)
+void renderScreenSpaceQuad(uint8_t _view, bgfx::ProgramHandle _program, float _x, float _y, float _width, float _height)
 {
 	bgfx::TransientVertexBuffer tvb;
 	bgfx::TransientIndexBuffer tib;
@@ -161,10 +161,10 @@ class ExampleRaymarch : public entry::AppI
 		if (!entry::processEvents(m_width, m_height, m_debug, m_reset) )
 		{
 			// Set view 0 default viewport.
-			bgfx::setViewRect(0, 0, 0, m_width, m_height);
+			bgfx::setViewRect(0, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 
 			// Set view 1 default viewport.
-			bgfx::setViewRect(1, 0, 0, m_width, m_height);
+			bgfx::setViewRect(1, 0, 0, uint16_t(m_width), uint16_t(m_height) );
 
 			// This dummy draw call is here to make sure that view 0 is cleared
 			// if no other draw calls are submitted to viewZ 0.
