+++ examples/common/entry/entry_windows.cpp
@@ -823,7 +823,7 @@ namespace entry
 		WindowHandle findHandle(HWND _hwnd)
 		{
 			bx::LwMutexScope scope(m_lock);
-			for (uint32_t ii = 0, num = m_windowAlloc.getNumHandles(); ii < num; ++ii)
+			for (uint16_t ii = 0, num = m_windowAlloc.getNumHandles(); ii < num; ++ii)
 			{
 				uint16_t idx = m_windowAlloc.getHandleAt(ii);
 				if (_hwnd == m_hwnd[idx])
